refactored notification cloud function
